Database commit
Huge improvement with the interaction of the queue and database.  Application will listen to config for decision of running as worker or as web interface, can be run as both.  Will update it's database entry accordingly.

Database entries for server information include CPU data, release data, total system memory.  Will include more optional data in the future to help the system better work with services.