Added socket.io authentication, setup for streams
Many things were changed (including some formatting improvements), the most notable of which is the addition of the `socket.io` package and the brief introduction of an authentication module to handle streams.  Packages are also now child of `lib/worker` rather than `lib/system`, as it is not planned for non-worker nodes to make use of package management in the future.